# docs/openapi/cxp-platform-api-v1.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "CXP Platform API",
    "description": "Civic Exchange Platform — citizen service-request intake, AI enrichment, governance discovery, and announcement feed. Authenticated endpoints require a Bearer JWT (Platform HS256 for operators, Supabase ES256 for citizens). Public endpoints are explicitly marked.\n\n## Compatibility & Deprecation Policy\n\nThe platform maintains backward-compatible aliases for one release cycle while clients migrate to canonical field names. All deprecated fields and endpoints are marked with a strikethrough in Swagger UI and carry `deprecated: true` in the generated OpenAPI schema — SDK generators will surface these as deprecated in generated client code.\n\n**Active transitional aliases (marked deprecated in schema):**\n- `ServiceRequestEvent.oldValue` → migrate to `oldStatus`\n- `ServiceRequestEvent.newValue` → migrate to `newStatus`\n- `ServiceRequestEvent.changedByRole` → migrate to `actorType`\n- `CitizenProfile.appUserId` → migrate to `id`\n- `CitizenProfile.postalCode` → migrate to `pinCode`\n- `UserProvision.appUserId` → migrate to `id`\n- `ServiceRequestCreate.aiConversationId` → not replaced; field is legacy from a discontinued conversational flow\n- Pagination `number` field → migrate to `page`\n\n**Deprecated endpoints:**\n- `POST /api/v1/ai/service-requests/analyze` → migrate to `POST /api/v1/service-requests/analyze` (identical contract)",
    "version": "v1"
  },
  "servers": [
    {
...
    }
...
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
    "/api/v1/users/provision": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Provision a citizen profile",
        "description": "Creates or updates the civic profile linked to the authenticated Supabase JWT. Identity (email) is extracted from the verified JWT claim — client-supplied identity fields are ignored. Safe to call on every app launch (upsert semantics). Requires: authenticated citizen session.",
        "operationId": "provision",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserProvisionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Validation failure or unresolvable wardId supplied in the request body.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected platform error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required — Bearer JWT is missing, expired, or invalid. Response body: {\"error\":\"Unauthorized\",\"message\":\"Invalid or missing token\"}"
          },
          "200": {
            "description": "Profile provisioned or updated",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UserProvisionResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/service-requests": {
      "get": {
        "tags": [
          "Service Requests"
        ],
        "summary": "List the authenticated citizen's service requests",
        "description": "Returns a paginated list of service requests submitted by the caller. Citizens see only their own requests. Default page size is 20; use ?page and ?size to paginate. Pagination envelope: 'page' is the canonical 0-based page index. 'number' is a deprecated backward-compatibility alias for 'page' — SDK consumers must migrate to 'page'; 'number' will be removed in a future release.",
        "operationId": "list",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected platform error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "200": {
            "description": "Paginated list of service requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedServiceRequestResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required — Bearer JWT is missing, expired, or invalid. Response body: {\"error\":\"Unauthorized\",\"message\":\"Invalid or missing token\"}"
          }
        }
      },
      "post": {
        "tags": [
          "Service Requests"
        ],
        "summary": "Submit a civic service request",
        "description": "Creates a new service request scoped to the authenticated citizen. Ward routing is derived server-side from the supplied GPS coordinates — no ward selection is required from the client. Supply an Idempotency-Key header to prevent duplicate submissions on retry.",
        "operationId": "create",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ServiceRequestCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Validation failure — required fields missing or invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected platform error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "Cannot assign ward jurisdiction — no GPS coordinates supplied and no ward in the citizen's profile. Complete onboarding (POST /api/v1/users/provision) or supply latitude/longitude in the request body.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required — Bearer JWT is missing, expired, or invalid. Response body: {\"error\":\"Unauthorized\",\"message\":\"Invalid or missing token\"}"
          },
          "200": {
            "description": "Service request created, or cached response returned for a duplicate Idempotency-Key.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceRequestResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/service-requests/analyze": {
      "post": {
        "tags": [
          "Service Requests"
        ],
        "summary": "Preview AI enrichment for a service request",
        "description": "Runs AI classification on the supplied title, description, and GPS coordinates. Returns a suggested category, priority, and routing hint — no data is persisted. Intended for pre-submit UX enrichment; the result is advisory, not authoritative.",
        "operationId": "analyze",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AiServiceRequestAnalyzeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Validation failure — required fields missing or invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected platform error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required — Bearer JWT is missing, expired, or invalid. Response body: {\"error\":\"Unauthorized\",\"message\":\"Invalid or missing token\"}"
          },
          "200": {
            "description": "AI enrichment preview returned",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AiServiceRequestAnalyzeResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/ai/service-requests/analyze": {
      "post": {
        "tags": [
          "AI"
        ],
        "summary": "AI service-request enrichment preview (deprecated — use /service-requests/analyze)",
        "description": "DEPRECATED. Canonical replacement: POST /api/v1/service-requests/analyze — identical request/response contract, stable path. This endpoint is a legacy alias retained for one release cycle during client migration. It will be removed in Wave 7. SDK consumers must migrate to the canonical path before that release.",
        "operationId": "analyze_1",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AiServiceRequestAnalyzeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Validation failure — required fields missing or invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected platform error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "200": {
            "description": "AI enrichment preview returned — same shape as /service-requests/analyze",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AiServiceRequestAnalyzeResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required — Bearer JWT is missing, expired, or invalid. Response body: {\"error\":\"Unauthorized\",\"message\":\"Invalid or missing token\"}"
          }
        },
        "deprecated": true
      }
    },
...
    "/api/v1/service-requests/{id}/status": {
      "patch": {
        "tags": [
          "Service Requests"
        ],
        "summary": "Update service request status",
        "description": "Advances the lifecycle state of a service request. Operator-only — requires the OPERATOR role on the platform JWT. Each transition is recorded as an immutable event in the request's event log.",
        "operationId": "updateStatus",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ServiceRequestStatusUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Validation failure or invalid status transition.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected platform error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required — Bearer JWT is missing, expired, or invalid. Response body: {\"error\":\"Unauthorized\",\"message\":\"Invalid or missing token\"}"
          },
          "200": {
            "description": "Service request status updated",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceRequestResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden — the caller does not have the OPERATOR role."
          },
          "404": {
            "description": "Service request not found for the given ID.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/service-requests/{id}": {
      "get": {
        "tags": [
          "Service Requests"
        ],
        "summary": "Retrieve a service request",
        "description": "Returns a single service request by ID. Citizens can retrieve only their own requests. Operators with a valid platform JWT can retrieve any request within their tenant.",
        "operationId": "get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected platform error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "200": {
            "description": "Service request returned",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceRequestResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required — Bearer JWT is missing, expired, or invalid. Response body: {\"error\":\"Unauthorized\",\"message\":\"Invalid or missing token\"}"
          },
          "404": {
            "description": "Service request not found for the given ID.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/service-requests/{id}/events": {
      "get": {
        "tags": [
          "Service Requests"
        ],
        "summary": "List the event history of a service request",
        "description": "Returns the immutable ordered event log for the given service request — status transitions, AI classifications, and external sync events. Events are append-only and never modified after creation.",
        "operationId": "getEvents",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected platform error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "200": {
            "description": "Ordered event log returned",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceRequestEventsResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required — Bearer JWT is missing, expired, or invalid. Response body: {\"error\":\"Unauthorized\",\"message\":\"Invalid or missing token\"}"
          },
          "404": {
            "description": "Service request not found for the given ID.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiError": {
        "type": "object",
        "properties": {
          "timestamp": {
            "type": "string",
            "description": "UTC timestamp of the error",
            "format": "date-time"
          },
          "status": {
            "type": "integer",
            "description": "HTTP status code",
            "format": "int32",
            "example": 400
          },
          "error": {
            "type": "string",
            "description": "Machine-readable error code, e.g. VALIDATION_FAILED, NOT_FOUND, BAD_REQUEST",
            "example": "VALIDATION_FAILED"
          },
          "message": {
            "type": "string",
            "description": "Human-readable error message"
          },
          "path": {
            "type": "string",
            "description": "Request path that produced the error",
            "example": "/api/v1/service-requests"
          },
          "fields": {
            "type": "object",
            "additionalProperties": {
              "type": "string",
              "description": "Field-level validation errors; present only on VALIDATION_FAILED responses",
              "nullable": true
            },
            "description": "Field-level validation errors; present only on VALIDATION_FAILED responses",
            "nullable": true
          }
        },
        "description": "Standard platform error response. The 'fields' map is present only on VALIDATION_FAILED (400) responses — it maps each failing field name to its constraint message. All other error types omit 'fields'."
      },
      "UserProvisionRequest": {
        "type": "object",
        "properties": {
          "fullName": {
            "type": "string",
            "description": "Citizen's full legal name.",
            "nullable": true
          },
          "mobileNumber": {
            "type": "string",
            "description": "Citizen's mobile number used for OTP authentication.",
            "nullable": true
          },
          "address": {
            "type": "string",
            "description": "Free-text address line for the citizen's registered location.",
            "nullable": true
          },
          "latitude": {
            "type": "number",
            "description": "GPS latitude for Path A ward resolution (highest-fidelity path). When provided alongside longitude, the backend resolves the ward automatically. Takes precedence over an explicit wardId.",
            "format": "double",
            "nullable": true
          },
          "longitude": {
            "type": "number",
            "description": "GPS longitude for Path A ward resolution. Same semantics as latitude.",
            "format": "double",
            "nullable": true
          },
          "pinCode": {
            "type": "string",
            "description": "Postal / PIN code (Indian convention). Also accepted as 'postalCode' for backward compatibility.",
            "nullable": true
          },
          "wardId": {
            "type": "string",
            "description": "Explicit ward identifier for Path B resolution — used when GPS is unavailable or the citizen manually corrects their ward via the governance discovery hierarchy. Must be a UUID from GET /governing-bodies/{id}/wards. Ignored when GPS coordinates are present and resolve successfully.",
            "format": "uuid",
            "nullable": true
          }
        },
        "description": "Citizen civic profile provisioning request for POST /api/v1/users/provision. Email is excluded — it is extracted from the validated JWT claims, making spoofed body values impossible. Ward resolution uses GPS coordinates (Path A, preferred) or an explicit wardId (Path B)."
      },
      "UserProvisionResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Canonical platform user identity (User.id). Stable across all profile updates. Persist this as the durable citizen identifier.",
            "format": "uuid"
          },
          "wardId": {
            "type": "string",
            "description": "Canonical ward identifier resolved for this citizen. Authoritative reference into GET /governing-bodies/{id}/wards.",
            "nullable": true
          },
          "wardName": {
            "type": "string",
            "description": "Denormalized ward display label. Use for display only; wardId is the authoritative reference.",
            "nullable": true
          },
          "zoneId": {
            "type": "string",
            "description": "Zone identifier within the ward hierarchy. Null when zone resolution is not configured.",
            "nullable": true
          },
          "governingBodyId": {
            "type": "string",
            "description": "Canonical governing authority identifier for this citizen's ward.",
            "nullable": true
          },
          "cityId": {
            "type": "string",
            "description": "Canonical city identifier resolved via the governance hierarchy.",
            "nullable": true
          },
          "city": {
            "type": "string",
            "description": "Denormalized city display label. Use for display only.",
            "nullable": true
          },
          "districtId": {
            "type": "string",
            "description": "Canonical district identifier resolved via the governance hierarchy.",
            "nullable": true
          },
          "district": {
            "type": "string",
            "description": "Denormalized district display label. Use for display only.",
            "nullable": true
          },
          "stateId": {
            "type": "string",
            "description": "Canonical state identifier resolved via the governance hierarchy.",
            "nullable": true
          },
          "state": {
            "type": "string",
            "description": "Denormalized state display label. Use for display only.",
            "nullable": true
          },
          "onboardingComplete": {
            "type": "boolean",
            "description": "True when the citizen has completed the onboarding flow (ward resolved)."
          },
          "appUserId": {
            "type": "string",
            "description": "DEPRECATED — use id. Backward-compatibility alias retained for one release cycle. Will be removed in a future release.",
            "format": "uuid",
            "deprecated": true
          }
        },
        "description": "Response from POST /api/v1/users/provision. Returned on both first-run provisioning and subsequent profile updates. The deprecated appUserId alias is emitted alongside id for backward compatibility."
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "Bearer JWT. Accepted formats: Platform HS256 (operator — carries governingBodyId claim) or Supabase ES256 (citizen — no tenantId claim). Header: Authorization: Bearer <token>",
        "in": "header",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}